# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_joe5saia_cautious-memory_internal_model.Field:
    properties:
      name:
        type: string
    type: object
  github_com_joe5saia_cautious-memory_internal_model.Profile:
    properties:
      email:
        type: string
      fields:
        items:
          $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Field'
        type: array
      firstName:
        type: string
      lastName:
        type: string
      phoneNumber:
        type: string
      school:
        type: string
      subfields:
        items:
          $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Subfield'
        type: array
    type: object
  github_com_joe5saia_cautious-memory_internal_model.Subfield:
    properties:
      name:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /fields:
    get:
      consumes:
      - application/json
      description: Get all fields
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Field'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all fields
      tags:
      - fields
    post:
      consumes:
      - application/json
      description: Create a field
      parameters:
      - description: Field
        in: body
        name: field
        required: true
        schema:
          $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Field'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Field'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a field
      tags:
      - fields
  /fields/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a field
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete a field
      tags:
      - fields
    get:
      consumes:
      - application/json
      description: Get a field by ID
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Field'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get a field
      tags:
      - fields
    post:
      consumes:
      - application/json
      description: Update a field
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: integer
      - description: Field
        in: body
        name: field
        required: true
        schema:
          $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Field'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update a field
      tags:
      - fields
  /profiles:
    get:
      consumes:
      - application/json
      description: Get all profiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Profile'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all profiles
      tags:
      - profiles
    post:
      consumes:
      - application/json
      description: Create a profile
      parameters:
      - description: Profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Profile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a profile
      tags:
      - profiles
  /profiles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete a profile
      tags:
      - profiles
    get:
      consumes:
      - application/json
      description: Get a profile by ID
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Profile'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get a profile
      tags:
      - profiles
    post:
      consumes:
      - application/json
      description: Update a profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Profile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update a profile
      tags:
      - profiles
  /subfields:
    get:
      consumes:
      - application/json
      description: Get all subfields
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Subfield'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all subfields
      tags:
      - subfields
    post:
      consumes:
      - application/json
      description: Create a subfield
      parameters:
      - description: Subfield
        in: body
        name: subfield
        required: true
        schema:
          $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Subfield'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Subfield'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a subfield
      tags:
      - subfields
  /subfields/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subfield
      parameters:
      - description: Subfield ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete a subfield
      tags:
      - subfields
    get:
      consumes:
      - application/json
      description: Get a subfield by ID
      parameters:
      - description: Subfield ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Subfield'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get a subfield
      tags:
      - subfields
    post:
      consumes:
      - application/json
      description: Update a subfield
      parameters:
      - description: Subfield ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subfield
        in: body
        name: subfield
        required: true
        schema:
          $ref: '#/definitions/github_com_joe5saia_cautious-memory_internal_model.Subfield'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update a subfield
      tags:
      - subfields
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"